Update create-restaurant-tables.php
Standalone Example of the plugin for adding and updating restaurants to the owners WordPress database. Woocommerce has a special "Shop Manger" user. Owner would add users and their email address for each restaurant as a Shop Manager, then use the plugin to associate the email address with a restaurant that they can then submit orders to. Restaurants and associated email addresses are displayed int the table underneath the form. Also displayed is a symbol indicating whether user is online or not. Owner can also update listings, either associating a different email address with a restaurant or changing the name of the restaurant, or delete a restaurant (user is still in wp database, just no longer associated with a restaurant). email addresses already associated with a restaurant are disabled.

To see the "plugin " in action: http://www.ronaldrferrucci.com/fooddelivery/create-restaurant-tables.php